external-recorder: ignore subsequent elements from base_data array
I think (but I'm not sure) that the array can contain several elements
even if there is only one eye camera, in case the world camera has a
lower FPS, so that the gaze data structure contains several frames from
the eye camera.

Let's ignore this for now, for the MRI-compatible camera there is anyway
no world camera, so I think for that setup the problem will not occur.